Main layout - players and scrolling grid
So the main layout is now working (yay) - almost… (boo)

We have two sections where player details live and these scale
depending on device which is nice. The grid lives in the middle between
them.

Only problem is that the grid isn’t scaling to the area - thats the
next step