Disabled the timed game and changed the checking for valid moves
Disabled the timer for the game as it is working.

The hand over to an other player still need to be tested throughly to
ensure my logic is correct. At the moment I am simply displaying
dialogue boxes if a user has no more moves available - no passing of
the turn has been configured yet.