the database update (Still bugs)
this is the latest database update, the database creates otherwise it
would crash on the first screen. it doesnt like adding or displaying i
believe these problems are due to the data types that are being inserted
and displayed. the facebook parts to this are the bare min as its just
the add into the gradle.  facebook developer has the rest of the
instructions